irqchip: mpm: Wire up irq_{en,dis}able and trigger parent
On IRQ enable and disable operations, the parent wasn't getting
triggered, or was getting called with the wrong function:
specifically, the irq_disable operation was wired up to the
msm_mpm_gic_chip_mask function, which was calling the relative
irq_mask handle on the interrupt parent, instead of irq_disable.

To address this issue, add specific enable and disable functions
in MPM, doing the same as mask and unmask (as that's what en/dis
means in MPM context), but calling the right handling on the
parent interrupt controller.

Of course, change the irq_disable handler of the mpm-gic to the
newly introduced specific function, instead of the mask one.

Signed-off-by: Raphiel Rollerscaperers <rapherion@raphielgang.org>
Signed-off-by: Twisted Prime <36546624+TwistedPrime@users.noreply.github.com>