pinctrl: msm: Remove irq_enable callback, always clear when unmasking
The irq_enable callback is not really necessary: the pins are
anyway enabled when we get to the pinctrl driver.

Also, following the previous versions of this driver, instead
of clearing the interrupt status bit only if the interrupt type
is TRIGGER_HIGH or TRIGGER_LOW, clear it for every type, as the
RISING/FALLING status might get stuck during some suspend cycle,
getting them unable to fire again to eventually trigger a wakeup.

Signed-off-by: Raphiel Rollerscaperers <rapherion@raphielgang.org>